modularized polyfills for web5
