+++ package.json
@@ -2,5 +2,9 @@
   "dependencies": {
     "@tbd54566975/dwn-sdk-js": "0.2.4",
     "@web5/api": "0.8.1"
+  },
+  "scripts": {
+    "test:web5": "pnpm -r --workspace-concurrency 1 --no-bail test:web5",
+    "test:dwn": "pnpm -r --workspace-concurrency 1 --no-bail test:dwn"
   }
 }
