+++ tests/node-cjs/dwn-sdk-test.js
@@ -1,3 +1,4 @@
+require("../util/node-polyfill");
 const {
   Dwn,
   DataStream,
