+++ tests/node-cjs/web5-test.js
@@ -1,3 +1,4 @@
+require("../util/node-polyfill");
 const { Web5 } = require("@web5/api");
 const checkWeb5 = require("../util/web5-test.js");
