+++ tests/node-esm/web5-test.js
@@ -1,3 +1,4 @@
+import "../util/node-polyfill.js";
 import { Web5 } from "@web5/api";
 import checkWeb5 from "../util/web5-test.js";
